fix: 에러 출력 문구 및 저장 코드 수정

+++ src/main/java/subway/service/StationManager.java
@@ -8,22 +8,35 @@ import subway.service.utils.Verify;
 
 import java.util.List;
 
+import static subway.controller.utils.Constants.STATION;
 import static subway.service.InitSubwayValues.stationRepo;
 
 // 역 관련 기능하는 서비스
 public class StationManager extends Verify implements Managerable {
+    InputException inputException = new InputException();
 
-    public void register(String name){
-        stationRepo.addStation(new Station(name));
+    public boolean register(String name){
+        if(isEmpty(STATION, name)) {
+            stationRepo.addStation(new Station(name));
+            return true ;
+        }
+        inputException.alreadyCreatedStation();
+        return false;
     }
     @Override
     public boolean delete(final String name) {
         Station station = stationRepo.getStationByName(name);
         List<Line> lines = station.getLine();
 
+        if(station.hasNoLine() == false) {
+            inputException.haveLines();
+            return false;
+        }
+
         for(Line line : lines){
             line.deleteStation(name);
         }
+
         return stationRepo.deleteStation(name);
     }
 
@@ -45,7 +58,6 @@ public class StationManager extends Verify implements Managerable {
     }
 
     public boolean haveSameName(String station){
-        if(haveSameNameStation(station)) return true;
-        return false;
+        return haveSameNameStation(station) ;
     }
 }
